Move auto_date_histogram aggregation to aggregations module.
Relates to #90283
