+++ modules/aggregations/src/main/java/org/elasticsearch/aggregations/AggregationsPlugin.java
@@ -10,6 +10,8 @@ package org.elasticsearch.aggregations;
 
 import org.elasticsearch.aggregations.bucket.AdjacencyMatrixAggregationBuilder;
 import org.elasticsearch.aggregations.bucket.InternalAdjacencyMatrix;
+import org.elasticsearch.aggregations.bucket.histogram.AutoDateHistogramAggregationBuilder;
+import org.elasticsearch.aggregations.bucket.histogram.InternalAutoDateHistogram;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.plugins.SearchPlugin;
 
@@ -24,7 +26,13 @@ public class AggregationsPlugin extends Plugin implements SearchPlugin {
                 AdjacencyMatrixAggregationBuilder.NAME,
                 AdjacencyMatrixAggregationBuilder::new,
                 AdjacencyMatrixAggregationBuilder::parse
-            ).addResultReader(InternalAdjacencyMatrix::new)
+            ).addResultReader(InternalAdjacencyMatrix::new),
+            new AggregationSpec(
+                AutoDateHistogramAggregationBuilder.NAME,
+                AutoDateHistogramAggregationBuilder::new,
+                AutoDateHistogramAggregationBuilder.PARSER
+            ).addResultReader(InternalAutoDateHistogram::new)
+                .setAggregatorRegistrar(AutoDateHistogramAggregationBuilder::registerAggregators)
         );
     }
 }
