+++ modules/aggregations/src/test/java/org/elasticsearch/aggregations/pipeline/SerialDifferenceTests.java
@@ -6,28 +6,30 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.search.aggregations.pipeline;
+package org.elasticsearch.aggregations.pipeline;
 
 import org.elasticsearch.search.aggregations.AggregationBuilder;
 import org.elasticsearch.search.aggregations.BasePipelineAggregationTestCase;
 import org.elasticsearch.search.aggregations.PipelineAggregationBuilder;
 import org.elasticsearch.search.aggregations.bucket.terms.TermsAggregationBuilder;
 import org.elasticsearch.search.aggregations.pipeline.BucketHelpers.GapPolicy;
+import org.elasticsearch.search.aggregations.pipeline.SerialDiffPipelineAggregationBuilder;
 
 import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
 
+import static java.util.Collections.emptyList;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.nullValue;
 
-public class DerivativeTests extends BasePipelineAggregationTestCase<DerivativePipelineAggregationBuilder> {
+public class SerialDifferenceTests extends BasePipelineAggregationTestCase<SerialDiffPipelineAggregationBuilder> {
 
     @Override
-    protected DerivativePipelineAggregationBuilder createTestAggregatorFactory() {
+    protected SerialDiffPipelineAggregationBuilder createTestAggregatorFactory() {
         String name = randomAlphaOfLengthBetween(3, 20);
         String bucketsPath = randomAlphaOfLengthBetween(3, 20);
-        DerivativePipelineAggregationBuilder factory = new DerivativePipelineAggregationBuilder(name, bucketsPath);
+        SerialDiffPipelineAggregationBuilder factory = new SerialDiffPipelineAggregationBuilder(name, bucketsPath);
         if (randomBoolean()) {
             factory.format(randomAlphaOfLengthBetween(1, 10));
         }
@@ -35,11 +37,7 @@ public class DerivativeTests extends BasePipelineAggregationTestCase<DerivativeP
             factory.gapPolicy(randomFrom(GapPolicy.values()));
         }
         if (randomBoolean()) {
-            if (randomBoolean()) {
-                factory.unit(String.valueOf(randomInt()));
-            } else {
-                factory.unit(String.valueOf(randomIntBetween(1, 10) + randomFrom("s", "m", "h", "d", "w", "M", "y")));
-            }
+            factory.lag(randomIntBetween(1, 1000));
         }
         return factory;
     }
@@ -51,26 +49,31 @@ public class DerivativeTests extends BasePipelineAggregationTestCase<DerivativeP
         assertThat(
             validate(
                 PipelineAggregationHelperTests.getRandomSequentiallyOrderedParentAgg(),
-                new DerivativePipelineAggregationBuilder("name", "valid")
+                new SerialDiffPipelineAggregationBuilder("name", "valid")
             ),
             nullValue()
         );
     }
 
-    /**
-     * The validation should throw an IllegalArgumentException, since parent
-     * aggregation is not a type of HistogramAggregatorFactory,
-     * DateHistogramAggregatorFactory or AutoDateHistogramAggregatorFactory.
-     */
-    public void testValidateException() throws IOException {
+    public void testInvalidParent() throws IOException {
         final Set<PipelineAggregationBuilder> aggBuilders = new HashSet<>();
-        aggBuilders.add(new DerivativePipelineAggregationBuilder("deriv", "der"));
-        AggregationBuilder parent = new TermsAggregationBuilder("name");
+        aggBuilders.add(createTestAggregatorFactory());
+        AggregationBuilder parent = new TermsAggregationBuilder("t");
         assertThat(
-            validate(parent, new DerivativePipelineAggregationBuilder("name", "invalid_agg>metric")),
+            validate(parent, new SerialDiffPipelineAggregationBuilder("name", "invalid_agg>metric")),
             equalTo(
-                "Validation Failed: 1: derivative aggregation [name] must have a histogram, "
+                "Validation Failed: 1: serial_diff aggregation [name] must have a histogram, "
                     + "date_histogram or auto_date_histogram as parent;"
+            )
+        );
+    }
+
+    public void testNoParent() {
+        assertThat(
+            validate(emptyList(), new SerialDiffPipelineAggregationBuilder("name", "invalid_agg>metric")),
+            equalTo(
+                "Validation Failed: 1: serial_diff aggregation [name] must have a histogram, "
+                    + "date_histogram or auto_date_histogram as parent but doesn't have a parent;"
             )
         );
     }
