+++ modules/aggregations/src/yamlRestTest/resources/rest-api-spec/test/aggregations/330_auto_date_histogram.yml
